+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @open-cells/core
 
+## 1.1.0-rc.0
+
+### Minor Changes
+
+- refactor core-plugin package
+
 ## 1.0.1
 
 ### Patch Changes