+++ packages/example/recipes-app/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @open-cells/recipes-app
 
+## 0.2.2-rc.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @open-cells/page-mixin@1.2.0-rc.0
+  - @open-cells/core@1.1.0-rc.0
+  - @open-cells/element-controller@1.0.2-rc.0
+  - @open-cells/page-controller@1.0.2-rc.0
+
 ## 0.2.1
 
 ### Patch Changes