+++ packages/page-mixin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @open-cells/page-mixin
 
+## 1.2.0-rc.0
+
+### Minor Changes
+
+- refactor core-plugin package
+
+### Patch Changes
+
+- Updated dependencies
+  - @open-cells/core-plugin@1.2.0-rc.0
+
 ## 1.1.1
 
 ### Patch Changes